refactor(server): extract discovery and sync routes
